+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+## v0.2.0
+
+- Changed iterator API: Removed `hasNext()`. Instead of throwing an error in
+  `next()` after the last item, `undefined` is returned.
+
 ## v0.1.0
 
 - Initial release